clean up

+++ src/main/java/com/tregouet/occam/alg/setters/differentiae_coeff/impl/NoCoeff.java
@@ -7,13 +7,13 @@ import com.tregouet.tree_finder.data.Tree;
 public class NoCoeff implements DifferentiaeCoeffSetter {
 
 	public static final NoCoeff INSTANCE = new NoCoeff();
-	
+
 	private NoCoeff() {
 	}
-	
+
 	@Override
 	public void accept(AbstractDifferentiae abstractDifferentiae) {
-		abstractDifferentiae.setWeightCoeff((double) 1);
+		abstractDifferentiae.setWeightCoeff(1);
 	}
 
 	@Override
