+++ src/main/java/com/tregouet/occam/alg/setters/weighs/properties/impl/BoundVarCounter.java
@@ -11,10 +11,10 @@ import com.tregouet.occam.data.representations.transitions.IApplication;
 public class BoundVarCounter implements PropertyWeigher {
 
 	public static final BoundVarCounter INSTANCE = new BoundVarCounter();
-	
+
 	private BoundVarCounter() {
 	}
-	
+
 	@Override
 	public void accept(IProperty property) {
 		Set<AVariable> varInstantiatedByNonRedundantApp = new HashSet<>();
@@ -23,7 +23,7 @@ public class BoundVarCounter implements PropertyWeigher {
 				varInstantiatedByNonRedundantApp.add(app.getInputConfiguration().getStackSymbol());
 			}
 		}
-		property.setWeight((double) varInstantiatedByNonRedundantApp.size());
+		property.setWeight(varInstantiatedByNonRedundantApp.size());
 	}
 
 }
